Restructure dynamic list component creation
